Rename TestkitTest2 to testkitTests
